requirements.txt
anthropic>=0.40.0
python-telegram-bot[job-queue]>=21.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.2.0
pytz>=2024.1
redis>=5.0.0
requests>=2.31.0
feedparser>=6.0.0
python-docx>=1.1.2
python-pptx>=1.0.2
openai>=1.100.0
pymupdf>=1.24.0
